backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# AI/LLM - HuggingFace only
huggingface-hub==0.20.3

# Document Parsing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
